refactor: move room crud in a separate component
